+++ libcnb-test/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [0.1.1] 2022-02-03
+
+- Connect to Docker demon via HTTP by default, only fall back to local socket/named pipe if no `DOCKER_HOST`
+environment variable is set. ([#305](https://github.com/Malax/libcnb.rs/pull/305))
+
 ## [0.1.0] 2022-02-02
 
 - Initial release ([#277](https://github.com/Malax/libcnb.rs/pull/277)).